+++ GameTest/Source/Scenery/Ground.h
@@ -7,11 +7,11 @@
 struct Ground
 {
 	Vector4 m_position;
-	int m_scale = 1;
+	float m_scale = 1.0f;
 	int m_width = 1;
 	int m_length = 1;
 
-	Ground(int scale, int width, int length);
+	Ground(float scale, int width, int length);
 	void GenerateTiles(ECS& ecs);
 	float GetMaxX();
 	float GetMaxZ();
